Update README.md
Most importantly with there was a misunderstanding on my part, the
algorithm isn't lock-free when consuming but its seems pretty good
anyway. It is lock free on the producer side which is most important
to me, but long term we'll see how it goes.

The performance numbers should be taken with a large grain of salt.
There is quite a bit of variability, 3x. I've seen this code take about
1.5minutes and also 30secs. The 30secs is seen after a cold boot of the
computer and I saw about 1.5minutes after a restart of the computer.

Also, the test only has the main thread running 100%, eventually I'll
probably change the client code to also generate messages to other
clients to try to saturate computer.