Change debug to output gTick, added USE_RMV added DELAY
Having gTick allows easier debugging to be able to reference
lines in the output. Another option would be to use Time Stamp Counter
but since its generally not synchronized across CPU's it isn't
very good for that purpose.

USE_RMV controls if we always use rmv or rmv_non_stalling. The
performance between the two looks to be about the same.

DELAY adds a usleep to add and returnto make it more likely that we
get a context switch at the critical spot.

Added VOLATILE in case that make a code generation difference, it doesn't
for NON Atomic types.