Account for msgs and cmds processed.
Now msgs_processed is the total number of messages removed from fifo's and
cmds_processed is the number of command messages processed by the
clients.

So cmds_processed represents a measure of "real" work by the system and
msgs_processed is a measure of time to add and remove a message.

Here are some numbers, not neccessarily accurate but should be useful
for detecting algorithm changes:
```
wink@wink-desktop:~/prgs/test-mpscfifo (master)
$ ./test 12 10000000 65536
test client_count=12 loops=10000000 msg_count=65536
     1 7facb820d700  multi_thread_msg:+client_count=12 loops=10000000 msg_count=65536
     2 7facb820d700  multi_thread_msg: msgs_processed=1611401675 msgs_sent=70210537 no_msgs_count=49789463
     3 7facb820d700  startup=0.044393
     4 7facb820d700  looping=28.210781
     5 7facb820d700  disconneting=0.120743
     6 7facb820d700  stopping=0.007948
     7 7facb820d700  complete=0.008225
     8 7facb820d700  processing=28.348s
     9 7facb820d700  cmds_per_sec=28407061
    10 7facb820d700  ns_per_cmd=35.2ns
    11 7facb820d700  msgs_per_sec=56844183
    12 7facb820d700  ns_per_msg=17.6ns
    13 7facb820d700  total=28.392
    14 7facb820d700  multi_thread_msg:-error=0

Success
```